create function to add all appliances to postgresql🎲

+++ server/internal/entity/appliance.go
@@ -8,22 +8,22 @@ type Appliance struct {
 	Priority bool
 	Location string
 	Power    int
-	Energy   int
+	Energy   float64
 }
 
 type ApplianceRequest struct {
-	Name     string `json:"name"`
-	Priority bool   `json:"priority"`
-	Location string `json:"location"`
-	Power    int    `json:"power"`
-	Energy   int    `json:"energy"`
+	Name     string  `json:"name"`
+	Priority bool    `json:"priority"`
+	Location string  `json:"location"`
+	Power    int     `json:"power"`
+	Energy   float64 `json:"energy"`
 }
 
 type ApplianceResponse struct {
-	ID       uint   `json:"id"`
-	Name     string `json:"name"`
-	Priority bool   `json:"priority"`
-	Location string `json:"location"`
-	Power    int    `json:"power"`
-	Energy   int    `json:"energy"`
+	ID       uint    `json:"id"`
+	Name     string  `json:"name"`
+	Priority bool    `json:"priority"`
+	Location string  `json:"location"`
+	Power    int     `json:"power"`
+	Energy   float64 `json:"energy"`
 }
